scripts/stick-install-ng.sh: set partition type to 0/empty for windows
.. because since 2016, windows 10 ignores the hidden flag or a foreign
partition type and offers to FORMAT any such USB partitions! 🔨😏⌛️

For this to work, the kernel and init ramdisk need to be extracted to
the (EFI) boot partition (using iso-read from libcdio-utils), as grub
will refuse to boot from partitions marked type 0 (empty).

The grub.cfg template is adjusted accordingly.