scripts/build-live-iso: bash 5.2 compatibility
Bash 5.2 "started to fully parse the content of the process substitution
in the normal parsing phase". "So, if you have an extglob in shell
functions, you need to turn on `extglob' outside the function to make
sure that the function is parsed with `extglob' being turned on".

( https://lists.gnu.org/archive/html/bug-bash/2023-02/msg00115.html )