Add app_blender_for_artists to all variants
"Bforartists is a fork of the popular open source 3D software Blender.
Developed by some 3D enthusiasts to make Blender even better.

The primary goal of the Bforartists fork is to deliver a refined
graphical UI and a better usability to Blender. This means a complete
switch in the usage philosophy away from the hotkey centered UX towards
a user friendly, discoverable and intuitive graphical UI.

Bforartists is fully compatible with the Blender files and addons. Every
feature that works in Blender should also work in Bforartists. Also,
every release is always up to date with the newest Blender version. The
toolset is the same. Where we differ the most is the GUI and the
handling of UX.

Try it out for free. You can switch from Blender to Bforartists to
Blender without problems. You won’t lose anything 🙂"

https://www.bforartists.de/download/