live-boot_hook_persistence_prepare: handle hotfixed empty partition
An empty partition will happen if the F2FS kernel module is not
available during ISO building so that the persistence partition can
neither be mounted nor the directories and persistence.conf files
created. That's exactly what this boot hook is supposed to handle.

Change the if condition so it looks for the actual conf files, not just
the presence of any directory - which might well have been created by
applying a hotfix.